Xpetra MatrixUtils: ETI
